Consolidate common E3V2 code
