+++ Marlin/src/inc/Conditionals_post.h
@@ -421,8 +421,9 @@
   #endif
 #endif
 
-#if EITHER(DWIN_CREALITY_LCD_ENHANCED, DWIN_CREALITY_LCD_JYERSUI)
+#if EITHER(HAS_DWIN_E3V2, IS_DWIN_MARLINUI)
   #define HAS_LCD_BRIGHTNESS 1
+  #define MAX_LCD_BRIGHTNESS 31
 #endif
 
 /**
