+++ Marlin/src/lcd/e3v2/common/dwin_font.h
@@ -21,14 +21,18 @@
  */
 #pragma once
 
-/********************************************************************************
- * @file     lcd/e3v2/jyersui/dwin_lcd.h
- * @brief    DWIN screen control functions
- ********************************************************************************/
-
-#include "../common/dwin_api.h"
-
-// Draw the degree (°) symbol
-// Color: color
-//  x/y: Upper-left coordinate of the first pixel
-void DWIN_Draw_DegreeSymbol(uint16_t Color, uint16_t x, uint16_t y);
+/**
+ * 3-.0：The font size, 0x00-0x09, corresponds to the font size below:
+ * 0x00=6*12   0x01=8*16   0x02=10*20  0x03=12*24  0x04=14*28
+ * 0x05=16*32  0x06=20*40  0x07=24*48  0x08=28*56  0x09=32*64
+ */
+#define font6x12  0x00
+#define font8x16  0x01
+#define font10x20 0x02
+#define font12x24 0x03
+#define font14x28 0x04
+#define font16x32 0x05
+#define font20x40 0x06
+#define font24x48 0x07
+#define font28x56 0x08
+#define font32x64 0x09
