+++ Marlin/src/lcd/e3v2/enhanced/dwinui.h
@@ -30,109 +30,11 @@
 
 #include "../../../core/types.h"
 #include "dwin_lcd.h"
-
-// ICON ID
-#define ICON                7 // Icon set file 7.ICO
-
-#define ICON_LOGO                  0
-#define ICON_Print_0               1
-#define ICON_Print_1               2
-#define ICON_Prepare_0             3
-#define ICON_Prepare_1             4
-#define ICON_Control_0             5
-#define ICON_Control_1             6
-#define ICON_Leveling_0            7
-#define ICON_Leveling_1            8
-#define ICON_HotendTemp            9
-#define ICON_BedTemp              10
-#define ICON_Speed                11
-#define ICON_Zoffset              12
-#define ICON_Back                 13
-#define ICON_File                 14
-#define ICON_PrintTime            15
-#define ICON_RemainTime           16
-#define ICON_Setup_0              17
-#define ICON_Setup_1              18
-#define ICON_Pause_0              19
-#define ICON_Pause_1              20
-#define ICON_Continue_0           21
-#define ICON_Continue_1           22
-#define ICON_Stop_0               23
-#define ICON_Stop_1               24
-#define ICON_Bar                  25
-#define ICON_More                 26
-
-#define ICON_Axis                 27
-#define ICON_CloseMotor           28
-#define ICON_Homing               29
-#define ICON_SetHome              30
-#define ICON_PLAPreheat           31
-#define ICON_ABSPreheat           32
-#define ICON_Cool                 33
-#define ICON_Language             34
-
-#define ICON_MoveX                35
-#define ICON_MoveY                36
-#define ICON_MoveZ                37
-#define ICON_Extruder             38
-
-#define ICON_Temperature          40
-#define ICON_Motion               41
-#define ICON_WriteEEPROM          42
-#define ICON_ReadEEPROM           43
-#define ICON_ResumeEEPROM         44
-#define ICON_Info                 45
-
-#define ICON_SetEndTemp           46
-#define ICON_SetBedTemp           47
-#define ICON_FanSpeed             48
-#define ICON_SetPLAPreheat        49
-#define ICON_SetABSPreheat        50
-
-#define ICON_MaxSpeed             51
-#define ICON_MaxAccelerated       52
-#define ICON_MaxJerk              53
-#define ICON_Step                 54
-#define ICON_PrintSize            55
-#define ICON_Version              56
-#define ICON_Contact              57
-#define ICON_StockConfiguration   58
-#define ICON_MaxSpeedX            59
-#define ICON_MaxSpeedY            60
-#define ICON_MaxSpeedZ            61
-#define ICON_MaxSpeedE            62
-#define ICON_MaxAccX              63
-#define ICON_MaxAccY              64
-#define ICON_MaxAccZ              65
-#define ICON_MaxAccE              66
-#define ICON_MaxSpeedJerkX        67
-#define ICON_MaxSpeedJerkY        68
-#define ICON_MaxSpeedJerkZ        69
-#define ICON_MaxSpeedJerkE        70
-#define ICON_StepX                71
-#define ICON_StepY                72
-#define ICON_StepZ                73
-#define ICON_StepE                74
-#define ICON_Setspeed             75
-#define ICON_SetZOffset           76
-#define ICON_Rectangle            77
-#define ICON_BLTouch              78
-#define ICON_TempTooLow           79
-#define ICON_AutoLeveling         80
-#define ICON_TempTooHigh          81
-#define ICON_NoTips_C             82
-#define ICON_NoTips_E             83
-#define ICON_Continue_C           84
-#define ICON_Continue_E           85
-#define ICON_Cancel_C             86
-#define ICON_Cancel_E             87
-#define ICON_Confirm_C            88
-#define ICON_Confirm_E            89
-#define ICON_Info_0               90
-#define ICON_Info_1               91
+#include "../common/dwin_set.h"
+#include "../common/dwin_font.h"
+#include "../common/dwin_color.h"
 
 // Extra Icons
-#define ICON_AdvSet               ICON_Language
 #define ICON_Brightness           ICON_Motion
 #define ICON_Cancel               ICON_StockConfiguration
 #define ICON_CustomPreheat        ICON_SetEndTemp
@@ -143,10 +45,6 @@
 #define ICON_FilSet               ICON_ResumeEEPROM
 #define ICON_FilUnload            ICON_ReadEEPROM
 #define ICON_Flow                 ICON_StepE
-#define ICON_HomeOffset           ICON_AdvSet
-#define ICON_HomeOffsetX          ICON_StepX
-#define ICON_HomeOffsetY          ICON_StepY
-#define ICON_HomeOffsetZ          ICON_StepZ
 #define ICON_LevBed               ICON_SetEndTemp
 #define ICON_Lock                 ICON_Cool
 #define ICON_ManualMesh           ICON_HotendTemp
@@ -154,13 +52,8 @@
 #define ICON_MeshSave             ICON_WriteEEPROM
 #define ICON_MoveZ0               ICON_HotendTemp
 #define ICON_Park                 ICON_Motion
-#define ICON_PIDbed               ICON_SetBedTemp
 #define ICON_PIDcycles            ICON_ResumeEEPROM
-#define ICON_PIDNozzle            ICON_SetEndTemp
 #define ICON_PIDValue             ICON_Contact
-#define ICON_ProbeOffsetX         ICON_StepX
-#define ICON_ProbeOffsetY         ICON_StepY
-#define ICON_ProbeOffsetZ         ICON_StepZ
 #define ICON_ProbeSet             ICON_SetEndTemp
 #define ICON_ProbeTest            ICON_SetEndTemp
 #define ICON_Pwrlossr             ICON_Motion
@@ -169,47 +62,6 @@
 #define ICON_Scolor               ICON_MaxSpeed
 #define ICON_SetCustomPreheat     ICON_SetEndTemp
 #define ICON_Sound                ICON_Cool
-
-/**
- * 3-.0：The font size, 0x00-0x09, corresponds to the font size below:
- * 0x00=6*12   0x01=8*16   0x02=10*20  0x03=12*24  0x04=14*28
- * 0x05=16*32  0x06=20*40  0x07=24*48  0x08=28*56  0x09=32*64
- */
-#define font6x12  0x00
-#define font8x16  0x01
-#define font10x20 0x02
-#define font12x24 0x03
-#define font14x28 0x04
-#define font16x32 0x05
-#define font20x40 0x06
-#define font24x48 0x07
-#define font28x56 0x08
-#define font32x64 0x09
-
-// Extended and default UI Colors
-#define RGB(R,G,B)  (R << 11) | (G << 5) | (B) // R,B: 0..31; G: 0..63
-#define GetRColor(color) ((color >> 11) & 0x1F)
-#define GetGColor(color) ((color >>  5) & 0x3F)
-#define GetBColor(color) ((color >>  0) & 0x1F)
-
-#define Color_White       0xFFFF
-#define Color_Bg_Window   0x31E8  // Popup background color
-#define Color_Bg_Blue     0x1125  // Dark blue background color
-#define Color_Bg_Black    0x0841  // Black background color
-#define Color_Bg_Red      0xF00F  // Red background color
-#define Popup_Text_Color  0xD6BA  // Popup font background color
-#define Line_Color        0x3A6A  // Split line color
-#define Rectangle_Color   0xEE2F  // Blue square cursor color
-#define Percent_Color     0xFE29  // Percentage color
-#define BarFill_Color     0x10E4  // Fill color of progress bar
-#define Select_Color      0x33BB  // Selected color
-
-#define Color_Black           0
-#define Color_Red             RGB(31,0,0)
-#define Color_Yellow          RGB(31,63,0)
-#define Color_Green           RGB(0,63,0)
-#define Color_Aqua            RGB(0,63,31)
-#define Color_Blue            RGB(0,0,31)
 
 // Default UI Colors
 #define Def_Background_Color  Color_Bg_Black
@@ -338,7 +190,7 @@ namespace DWINUI {
   extern void (*onMenuDraw)(MenuClass* menu);
 
   // DWIN LCD Initialization
-  void Init(void);
+  void Init();
 
   // Set text/number font
   void SetFont(uint8_t cfont);
