+++ buildroot/tests/STM32F103RET6_creality
@@ -10,21 +10,21 @@ set -e
 # Build with configs included in the PR
 #
 use_example_configs "Creality/Ender-3 V2/CrealityUI"
-opt_enable MARLIN_DEV_MODE BUFFER_MONITORING
+opt_enable MARLIN_DEV_MODE BUFFER_MONITORING BLTOUCH AUTO_BED_LEVELING_BILINEAR Z_SAFE_HOMING
 exec_test $1 $2 "Ender 3 v2 with CrealityUI" "$3"
 
 use_example_configs "Creality/Ender-3 V2/CrealityUI"
 opt_disable DWIN_CREALITY_LCD
-opt_enable DWIN_CREALITY_LCD_ENHANCED
+opt_enable DWIN_CREALITY_LCD_ENHANCED BLTOUCH AUTO_BED_LEVELING_UBL Z_SAFE_HOMING
 exec_test $1 $2 "Ender 3 v2 with Enhanced UI" "$3"
 
 use_example_configs "Creality/Ender-3 V2/CrealityUI"
 opt_disable DWIN_CREALITY_LCD
-opt_enable DWIN_CREALITY_LCD_JYERSUI
+opt_enable DWIN_CREALITY_LCD_JYERSUI AUTO_BED_LEVELING_BILINEAR PROBE_MANUALLY
 exec_test $1 $2 "Ender 3 v2 with JyersUI" "$3"
 
 use_example_configs "Creality/Ender-3 V2/MarlinUI"
-opt_add SDCARD_EEPROM_EMULATION
+opt_add SDCARD_EEPROM_EMULATION NOZZLE_AS_PROBE AUTO_BED_LEVELING_BILINEAR Z_SAFE_HOMING
 exec_test $1 $2 "Ender 3 v2 with MarlinUI" "$3"
 
 restore_configs
