+++ ini/features.ini
@@ -44,10 +44,11 @@ HAS_SPI_TFT                            = src_filter=+<src/HAL/STM32/tft/tft_spi.
 I2C_EEPROM                             = src_filter=+<src/HAL/shared/eeprom_if_i2c.cpp>
 SOFT_I2C_EEPROM                        = SlowSoftI2CMaster, SlowSoftWire=https://github.com/felias-fogg/SlowSoftWire/archive/master.zip
 SPI_EEPROM                             = src_filter=+<src/HAL/shared/eeprom_if_spi.cpp>
+HAS_DWIN_E3V2|IS_DWIN_MARLINUI         = src_filter=+<src/lcd/e3v2/common>
 DWIN_CREALITY_LCD                      = src_filter=+<src/lcd/e3v2/creality>
 DWIN_CREALITY_LCD_ENHANCED             = src_filter=+<src/lcd/e3v2/enhanced>
 DWIN_CREALITY_LCD_JYERSUI              = src_filter=+<src/lcd/e3v2/jyersui>
-DWIN_MARLINUI_.+                       = src_filter=+<src/lcd/e3v2/marlinui>
+IS_DWIN_MARLINUI                       = src_filter=+<src/lcd/e3v2/marlinui>
 HAS_GRAPHICAL_TFT                      = src_filter=+<src/lcd/tft>
 IS_TFTGLCD_PANEL                       = src_filter=+<src/lcd/TFTGLCD>
 HAS_TOUCH_BUTTONS                      = src_filter=+<src/lcd/touch/touch_buttons.cpp>
